refactor: move middlewares
